🐛 [disasm-lib] Make the decoder clean under /W4 and GCC's strict warnings
Most of what the stricter warning levels reported was noise with one root
cause, so most fixes land in one place rather than at every site.

Text output went through _snprintf, deprecated and wrong on truncation: it
returns -1 and leaves the buffer unterminated, and the result was added to
the write position anyway. One AppendFormat() helper now formats with
vsnprintf, truncates at the buffer's end and keeps the terminator; GCC checks
its format strings. The MSVC-only %I64 conversions became the <inttypes.h>
macros, which also let GCC check them, and that exposed two diagnostics that
passed an integer where the message printed a string.

Pointers are widened to U64 through ULONG_PTR. On 32-bit GCC a direct cast
sign-extends, so an address above 2 GB became 0xFFFFFFFF8xxxxxxx; MSVC always
zero-extended.

The OPTYPE_REG_* macros gain parentheses, which silences 111 reports in the
opcode tables at once. Fourteen register-name tables nothing referenced are
deleted. ITYPE_EXT_UNUSED5 was 1 << 31, a signed overflow, and is now
unsigned. Two switches gain the default they fell through to anyway, dead
code after a break in AMODE_V goes, an assertion that a WORD is below
0x10000 goes, and the cmc opcode entry gets the slot it was missing; its
values do not change.

For -Wconversion and -Wsign-conversion, values stored in narrow bit-fields
are computed in unsigned arithmetic with the mask last, which is the only
form GCC accepts as fitting, and the address arithmetic that wraps on
purpose converts its signed operand explicitly.

With sanity checks disabled, assert() expanded to nothing, which left
variables unused and if statements without a body. It now names its
argument inside sizeof, so nothing is evaluated.

Two MSVC reports stay, suppressed around the one structure each concerns:
C4201 for the anonymous unions, which are standard C11 and C++, and C4324 for
the padding that U128's alignment asks for. The file-wide pragmas that
disabled C4996, C4311 and C4312 are gone.

+++ src/disasm-lib/src/misc.c
@@ -105,7 +105,7 @@ BYTE* HexToBinary(char* Input, DWORD InputLength, DWORD* OutputLength)
             if (i >= InputLength)
                 break;
 
-            if (!IsHexChar(Input[i]))
+            if (!IsHexChar((BYTE)Input[i]))
             {
                 // fprintf(stderr, "ERROR: invalid hex character at offset %lu (0x%04x)\n", i, i);
                 goto abort;
@@ -162,7 +162,7 @@ BYTE* HexToBinary(char* Input, DWORD InputLength, DWORD* OutputLength)
             }
             i += 2;
 
-            if (!IsHexChar(Input[i]))
+            if (!IsHexChar((BYTE)Input[i]))
             {
                 // fprintf(stderr, "ERROR: invalid hex character at offset %lu (0x%04x)\n", i, i);
                 goto abort;
@@ -203,7 +203,7 @@ BYTE* HexToBinary(char* Input, DWORD InputLength, DWORD* OutputLength)
     {
         for (i = 0; i < InputLength; i += 2)
         {
-            if (!IsHexChar(Input[i]))
+            if (!IsHexChar((BYTE)Input[i]))
             {
                 // fprintf(stderr, "ERROR: invalid hex character at offset %lu (0x%04x)\n", i, i);
                 goto abort;
